ADD style.ts for 코드 간결
간결함을 위한 css code는 style.ts에 생성
