--- src/components/container/index.tsx
@@ -1,18 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-
-interface ContainerProps {
-    children: React.ReactNode;
-}
-const ContainerWrapper = styled.main`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin: 64px auto;
-    min-width: 200px;
-    width: 768px;
-`
-
-const Container = ({children}: ContainerProps) => <ContainerWrapper>{children}</ContainerWrapper>
-
-export default Container;
