+++ src/components/index.ts
@@ -1,0 +1,15 @@
+// atoms
+export { default as Input } from "./atoms/input";
+export { default as ToggleTheme } from "./atoms/toggleTheme";
+
+// molecules
+export { default as Todo } from "./molecules/todo";
+
+// organisms
+export { default as Container } from "./organisms/container";
+export { default as TodoGroup } from "./organisms/todoGroup";
+
+// templates
+export { default as Footer } from "./templates/footer";
+export { default as Header } from "./templates/header";
+export { default as Navbar } from "./templates/navbar";
