+++ src/components/molecules/todo/index.tsx
@@ -1,6 +1,5 @@
 import { TodoTypes } from "../../../types";
-import styled from "styled-components";
-import Button from "../../atoms/button";
+import { ButtonStyle, CheckedStyle, TextStyle, TodoStyle } from "./style";
 
 interface TodoProps {
   todo: TodoTypes;
@@ -8,32 +7,16 @@ interface TodoProps {
   isChecked: (id: string) => void;
 }
 
-interface TextProps{
-  checked: boolean;
+interface ButtonProps {
+  children: string;
+  onClick: () => void;
 }
 
-const TodoStyle = styled.article`
-  width: 400px;
-  font-size: 24px;
-  margin-bottom: 16px;
-  justify-content: center;
-  white-space: nowrap;
-  overflow: auto;
-  min-width: 0;
-`;
 
-const CheckedStyle = styled.input`
-  width: 20px;
-  height: 20px;
-  display: inline;
-`
+const Button = ({ children, onClick }: ButtonProps) => {
+  return <ButtonStyle onClick={onClick}>{children}</ButtonStyle>
+};
 
-const TextStyle = styled.span<TextProps>`
-  display: inline;
-  margin-left: 8px;
-  text-decoration: ${(props) => props.checked && "line-through"};
-  color: ${(props) => props.checked && "gray"};
-`;
 
 const Todo = ({ todo, onRemove, isChecked }: TodoProps) => {
   return (
