+++ src/components/organisms/container/style.ts
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const ContainerWrapper = styled.main`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 16px auto;
+    min-width: 200px;
+    width: 768px;
+`
