+++ src/components/templates/footer/style.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const FooterStyle = styled.footer`
+    margin-top: -8px;
+    padding: 0px 8px;
+    color: ${(props) => props.theme.footerTextColor};
+`
