+++ src/components/templates/header/style.ts
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+
+export const HeaderWrapper = styled.h1`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  font-size: 46px;
+  opacity: 87%;
+`;
+
+export const DateStyle = styled.p`
+  font-size: 16px;
+  opacity: 87%;
+  margin-top: 4px;
+  margin-bottom: 32px;
+`;
